chore: setup ts, eslint and prettier
