+++ package.json
@@ -2,18 +2,25 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "typecheck": "tsc",
+        "format": "prettier --write \"./**/*.ts\"",
+        "lint": "eslint \"./**/*.ts\" --fix",
+        "start:dev": "tsnd --respawn --transpile-only ./index.ts",
+        "start": "node ./dist/index.js"
     },
-    "author": "Andrei Auchynnikau",
+    "author": "Ilja Gusiatin",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@typescript-eslint/eslint-plugin": "^5.61.0",
+        "eslint": "^8.44.0",
+        "eslint-config-prettier": "^8.8.0",
+        "eslint-plugin-prettier": "^4.2.1",
+        "nodemon": "^2.0.16",
+        "prettier": "^2.8.8",
+        "ts-node-dev": "^2.0.0",
+        "typescript": "^5.1.6"
     },
     "dependencies": {
         "ws": "^8.8.0"
